refactor(net): split stack adapter
